pyproject.toml
[tool.poetry]
name = "humber"
version = "0.1.0"
description = "An opinionated photo stream site generator"
authors = ["Aaron Toth <atoth89@gmail.com>"]
license = "Apache-2.0"
readme = "README.rst"
repository = "https://github.com/aaront/humber"
keywords = ["static", "generator", "photography", "gallery", "blog"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Console",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules"
]


[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^8.1.0"
Jinja2 = "^2.11.3"
click = "^7.1.2"
tomlkit = "^0.7.0"
python-frontmatter = "^0.5.0"
mistune = {version = "2.0.0a6", allow-prereleases = true}

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
flake8 = "^3.8.4"
flake8-black = "^0.2.1"
bandit = "^1.7.0"
mypy = "^0.800"

[tool.poetry.scripts]
humber = "humber.console:run"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
